+++ src/Enum/MeasurementUnit.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Janisvepris\ZplBuilder\Enum;
+
+enum MeasurementUnit: string
+{
+    case Dots = 'D';
+    case Inches = 'I';
+    case Millimeters = 'M';
+}